Changed a lot in MediaCore. Low level MediaReader interface changed.
1. Unifies the position/timestamp type used for MediaReader apis related
to ReadVideo/Audio, Seek operations. Changed the type from 'double'
which represent pos/time as a float value in second, to 'int64_t' which
represent pos/time as a integer value in millisecond. This unifies the
pos/time representation of MediaReader apis as the same of that in
'MultiTrack/Track/Clip' layer.
2. Improved the seek operations invoked times, reduce the redundantly
invoked seek to the low level MediaReader layer.
3. Fixed a bug in Overview that some snapshot can not be generated.